+++ CHANGELOG.md
@@ -70,12 +70,10 @@ This PR contains a series of PRs on multi-staking support and BTC staking integr
 - [#525](https://github.com/babylonlabs-io/babylon/pull/525) fix: add back `NewIBCHeaderDecorator` post handler
 - [#667](https://github.com/babylonlabs-io/babylon/pull/667) crypto: enable groupcheck in BLS verification/aggregation
 - [#660](https://github.com/babylonlabs-io/babylon/pull/660) fix: ecdsa verification
-<<<<<<< HEAD
 - [#680](https://github.com/babylonlabs-io/babylon/pull/680) crypto: fix bls rogue attack
-=======
 - [#673](https://github.com/babylonlabs-io/babylon/pull/673) fix: move bip322 signing
 functions to `testutil`
->>>>>>> main
+- [#683](https://github.com/babylonlabs-io/babylon/pull/683) crypto: fix eots signing timing attack
 
 ## v1.0.0-rc7